fix(v7.0/W6.3.A): address Copilot iter-2 findings (9)
Config type-stability:

- `tool_calling.max_iterations` + `handshake.ttl_seconds` were
  read via `env()` without casting, which can return strings.
  Downstream callers pass these into `max()` / loop counters.
  Now `(int) env(...)` to match `config/mcp.php`.

Model docblock honesty:

- The class doc said package writes "MUST pass
  `input_json_redacted`" but the previous iter changed the
  `creating()` hook to default it to `[]` when missing. Doc
  rewritten to match: package writes that omit the payload have
  the column auto-defaulted; writes that DO pass a payload are
  preserved verbatim.

Tenant scoping via trait scope:

- `EloquentMcpServerRegistry` was using a bare
  `where('tenant_id', ...)` even though `McpServer` carries the
  `BelongsToTenant` trait that provides `scopeForTenant()`.
  Switched both queries (`forTenant()` + `find()`) to use the
  trait scope — table-qualified column survives future joins
  without ambiguous-column errors.

Stdio command-line parser — handle quoted args properly:

- `preg_split('/\s+/')` shredded `--prefix="hello world"` into 3
  tokens, which would invoke the wrong upstream args. Replaced
  with a small state-machine tokenizer
  (`McpServerAdapter::tokeniseCommandLine()`) that:
    - keeps quoted regions together (`--prefix="hello world"` →
      one token `--prefix=hello world`);
    - strips the surrounding `"` characters;
    - drops POSIX-shell complexity (backslash-escapes, single
      quotes, env-substitution) — operators who need them should
      use a future v8 `command` + `args` split storage.
- New `test_stdio_transport_config_preserves_quoted_arguments`
  regression covers the three shapes: bare tokens, mid-token
  quotes, whole-token quotes.

HostBridge — omit empty tools array:

- `chat()` always set `$options['tools']` even when the catalog
  was empty. Sending `"tools": []` to OpenAI / OpenRouter
  behaves differently than omitting the field (and some
  providers reject empty arrays). Now the bridge passes the
  request through as a plain completion when the catalog is
  empty.
- New `test_chat_omits_tools_when_catalog_is_empty` regression.

Doc + test cleanups:

- `EloquentMcpServerRegistryTest` file-level docblock claimed
  "null tenant = platform-global only" — the actual behavior
  (after iter-1 fix) is "fall back to current TenantContext".
  Doc rewritten to match.
- `User::create()` fixtures in 3 V7 test files passed
  `tenant_id` even though `users` is intentionally not
  tenant-scoped in this repo — the column doesn't exist and
  `User::$fillable` doesn't include it. Dropped from all 3 to
  avoid the misleading no-op (and a future
  `MassAssignmentException` if strict mode lands).

Tests: 41 V7 / 101 assertions; 99 MCP-filtered / 266 assertions
— all green.

+++ config/mcp-pack.php
@@ -23,7 +23,10 @@ return [
     */
     'tool_calling' => [
         'enabled' => env('AI_AGENTIC_ENABLED', false),
-        'max_iterations' => env('MCP_PACK_TOOL_CALLING_MAX_ITERATIONS', 3),
+        // Cast to int — env() returns strings, and downstream consumers
+        // pass these straight into `max()` / loop counters. Matches the
+        // strict-int treatment in `config/mcp.php`.
+        'max_iterations' => (int) env('MCP_PACK_TOOL_CALLING_MAX_ITERATIONS', 3),
         'default_tool_choice' => env('MCP_PACK_TOOL_CHOICE', 'auto'),
     ],
 
@@ -33,7 +36,7 @@ return [
     |--------------------------------------------------------------------------
     */
     'handshake' => [
-        'ttl_seconds' => env('MCP_PACK_HANDSHAKE_TTL', 300),
+        'ttl_seconds' => (int) env('MCP_PACK_HANDSHAKE_TTL', 300),
     ],
 
     /*
